refactor(viewer): split the symbolic module and stop swallowing the debug-flag read (#2183)
Local CodeRabbit review of the extraction (the hosted one is rate limited).
Three of its four findings were worth acting on, all against code this PR
moved rather than wrote.

The new module landed at 453 lines, over the ~400 house limit. Split into
symbolic-shapes.ts (result contracts + the pure tessellation helpers, 198
lines) and symbolic-parse.ts (the WASM walk, 288). symbolic-parse re-exports
the shapes so consumers keep one import site. The walk is still byte-identical
to origin/main after the split, re-verified by diff.

debugEnabled swallowed a localStorage failure with a bare catch, which
AGENTS.md forbids. It now warns. Storage access does throw in some privacy
modes, so the fallback stays.

Adds tests for the one thing the extraction actually introduced,
createEmptyParseResult: every bucket present, all empty, and a FRESH result
per call. That last one matters because callers push into these arrays, so a
shared instance would leak one model's annotations into another's result.
Mutation-checked by returning a module-level singleton.

NOT fixed here, on purpose: CodeRabbit's remaining finding is that the walk
treats worldY === 0 as "no elevation" and falls back to the storey table, so a
primitive genuinely at zero cannot be distinguished from a missing one. That
is real, but it is pre-existing behaviour and changing it is not a move. It
wants its own issue and its own fixture.
